refactor: all queries in index, named to constants
